de: Generalize `Deserialize` implementation to use `ReadDoc`
While being tasked to use `AutoCommit` in our `automerge 0.5` upgrade
(because this functionality used/appeared to be built in to `Automerge`
on `automerge 0.2`), the serializer works naturally because it is
implemented for the generic `Transactable` but deserializing is only
implemented for `Automerge`.

Instead, implement the deserializer for `trait ReadDoc` so that both
`Automerge` and `AutoCommit` can be used.